add objetc pooling+ scriptableobject for baron character

+++ Assets/Scripts/ScriptableObject/CharacterStats.cs
@@ -7,7 +7,7 @@ public class CharacterStats : ScriptableObject
     [Header("Base stats")]
     public float maxHp = 36f;
     public float maxSpeed = 36f;
-    public float mana = 36f;
+    public int mana = 36;
 
     [Header("Combat")]
     public float maxDamage = 36f;
